requirements.txt
alembic==1.15.1
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fonttools==4.55.2
greenlet==3.1.1
kiwisolver==1.4.7
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
passlib[bcrypt]==1.7.4
pillow==11.0.0
plotly==5.24.1
pymongo==4.11.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
qrcode==8.1
six==1.17.0
SQLAlchemy==2.0.23
tenacity==9.0.0
typing_extensions==4.12.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==1.10.13
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
